Refactor character creation form into multiple components
